# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Gofin API
  description: >
    Personal finance management API.
    All protected endpoints require a JWT bearer token in the Authorization header.
    All ID fields use UUID v7 format.
  version: "1.0.0"
  contact:
    name: Gofin
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Transaction ID
...
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        title:
          type: string
          example: Bad Request
        detail:
          type: string
          example: Invalid request body
...
    GroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
...
    TransactionRequest:
      type: object
      required:
        - type
        - date
        - amount
        - source_id
      properties:
        type:
          type: string
          enum: [withdrawal, deposit, transfer]
        date:
          type: string
          format: date-time
        amount:
          type: string
        source_id:
          type: string
          format: uuid
        destination_id:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    SplitTransactionRequest:
      type: object
      required:
        - transactions
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionRequest"
...
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/groups:
    get:
      summary: List groups
      description: List all user groups the authenticated user belongs to.
      operationId: listGroups
      tags: [Groups]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of groups
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create group
      description: Create a new user group.
      operationId: createGroup
      tags: [Groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupRequest"
      responses:
        "201":
          description: Group created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/transactions:
    get:
      summary: List transactions
      description: List transactions with optional filtering and pagination.
      operationId: listTransactions
      tags: [Transactions]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: type
          in: query
          description: Filter by transaction type
          schema:
            type: string
            enum: [withdrawal, deposit, transfer]
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: account_id
          in: query
          schema:
            type: string
            format: uuid
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of transactions
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      summary: Create transaction
      description: Create a new double-entry transaction. Requires manage_transactions role.
      operationId: createTransaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "201":
          description: Transaction created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/transactions/split:
    post:
      summary: Create split transaction
      description: Create a split transaction with multiple journals. Requires manage_transactions role.
      operationId: createSplitTransaction
      tags: [Transactions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitTransactionRequest"
      responses:
        "201":
          description: Split transaction created
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/transactions/{id}:
    get:
      summary: Get transaction
      description: Get transaction details by ID.
      operationId: getTransaction
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction details
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update transaction
      description: Update a transaction. Requires manage_transactions role.
      operationId: updateTransaction
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionRequest"
      responses:
        "200":
          description: Transaction updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete transaction
      description: Delete a transaction. Requires manage_transactions role.
      operationId: deleteTransaction
      tags: [Transactions]
      parameters:
        - $ref: "#/components/parameters/TransactionId"
      responses:
        "200":
          description: Transaction deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # -- Categories -----------------------------------------------------------
...